+++ templates/admin/tab-dashboard.php
@@ -9,67 +9,6 @@ use AnyComment\Admin\AnyCommentStatistics;
 
 <div class="anycomment-tab anycomment-dashboard__tab" id="anycomment-dashboard">
 
-    <?php if (AnyCommentLinkHelper::get_language() === 'ru'): ?>
-        <div class="anycomment-dashboard__splitter">
-            <div class="grid-x">
-                <div class="small-12">
-                    <br>
-                    <div class="callout success">
-                        <h3>AnyComment теперь и <a href="https://anycomment.io" target="_blank">в облаке</a>!</h3>
-                        <p>Теперь AnyComment будет в двух версиях:</p>
-                        <ul>
-                            <li>1) Плагин, который после установки работает локально на вашем сайте.</li>
-                            <li>2) Современное, облачное SaaS решение, которое работает на наших серверах.</li>
-                        </ul>
-                        <p><strong>Почему стиот перейти на облачное решение?</strong></p>
-                        <ul>
-                            <li>- В х2 раза быстрее</li>
-                            <li>- Качественная поддержка</li>
-                            <li>- Никакой нагрузки на ваш сайт</li>
-                            <li>- Бесплатный тариф + адекватная стоимость</li>
-                        </ul>
-
-                        <p><a class="button" href="https://anycomment.io" target="_blank">Подробнее</a></p>
-
-                        <h4>Статус комментариев:</h4>
-
-                        <?php
-
-                        if (AnyCommentServiceApi::is_ready()):
-                            $syncInfo = AnyCommentServiceSyncCron::getSyncInfo();
-
-                            if ($syncInfo['complete_percent'] === 100): ?>
-                                <p>Все комментарии синхронизированы!</p>
-                            <?php else: ?>
-                                <p><?php echo sprintf('Обработано %s из %s комментариев, завершено на %s%%', $syncInfo['current'], $syncInfo['total'], $syncInfo['complete_percent']) ?></p>
-                            <?php endif; ?>
-
-                        <?php else: ?>
-                            <p>У вас еще не настроены синхронизация с сервисом. Для этого вам нужно:</p>
-                            <ul>
-                                <li>- Добавить свой сайт на <a
-                                            href="<?php echo AnyCommentLinkHelper::get_service_website() ?>"
-                                            target="_blank">сервисе</a>
-                                </li>
-                                <li>- Зайти на страницу сайта, под заголовком "Настройка синхронизации" нажмите на
-                                    кнопку "Настроить"
-                                </li>
-                                <li>- Далее следуйте инструкции на странице "Синхронизация"</li>
-                                <li>- После завершения вместо этого сообщения должна появиться
-                                    информация по синхронизации комментариев
-                                </li>
-                            </ul>
-
-                            <p>Если у вас возникли вопросы, пишите на support@anycomment.io. Мы постараемся помочь вам
-                                как можно быстрее!</p>
-                        <?php endif; ?>
-                    </div>
-                </div>
-
-            </div>
-        </div>
-    <?php endif; ?>
-
     <div class="grid-x anycomment-dashboard__splitter">
         <div class="cell large-6 medium-12 small-12 anycomment-dashboard__splitter-half anycomment-dashboard__splitter-half-commentators">
             <div class="grid-x align-center">
@@ -104,7 +43,7 @@ use AnyComment\Admin\AnyCommentStatistics;
             <?php
 
             $comments = AnyCommentStatistics::get_comment_data();
-            $users = AnyCommentStatistics::get_commentor_data();
+            $users    = AnyCommentStatistics::get_commentor_data();
             ?>
             <canvas id="anycomment-dashboard-chart"></canvas>
 
@@ -180,7 +119,7 @@ use AnyComment\Admin\AnyCommentStatistics;
             <?php
             $users = AnyCommentStatistics::get_most_active_users();
 
-            if (!empty($users)): ?>
+            if ( ! empty($users)): ?>
                 <ul>
                     <?php foreach ($users as $user): ?>
                         <li> <span class="anycomment-dashboard__statistics-userlist-avatar"