+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Freemius was cut-out. This removes problems for some websites + speeds-up plugin a bit.
 - Added support for BuddyPress, when somebody send "@" + username in comment. This user would be mentioned, #422
+- Moved AnyComment Cloud message to Integration tab, improved UI/UX of synchronization logic 
 
 **Fixes:**